diva_cloud/diva_django/requirements.txt
django==2.2.6
djangorestframework==3.10.3
six==1.12
python-dateutil==2.7.3
pyyaml==5.1.2
uritemplate==3.0.0
psycopg2>=2.7,<3.0
celery==4.4.0
redis==3.3.11
eventlet==0.26.1
numpy==1.18.3
pandas==1.0.3
matplotlib==3.2.1
sklearn==0.0
optuna==1.3.0
scikit-learn==0.21.3
xgboost==0.90
catboost==0.24.2
